requirements.txt
python-dotenv==1.0.0
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
openai==1.3.0